Add Ticket Library on SelectRutas.kt and SelectTarifa.kt
Added ticket library implementation on SelectRutas.kt and SelectTarifa.kt using the EscPOS

+++ app/src/main/java/com/smartgeeks/busticket/repository/ticket/TicketRepositoryImpl.kt
@@ -52,7 +52,7 @@ class TicketRepositoryImpl @Inject constructor(
         return if (InternetChecker.checkForInternetConnection(context)){
             Log.e("Repo saveTicket", "Internet: ")
             val response = ticketApi.saveTicket(ticketEntity)
-            if (response.estado == 1){
+            if (response.estado == 1 && ticketEntity.id > 0){
                 Log.e(TAG, "Deleting ticket: ${ticketEntity.id}")
                 ticketDAO.delete(ticketEntity)
             }
